implement editor v1 you need to change the lyrics in the file itself but editing notes themselves works in the timeline

+++ pTyping/Songs/Note.cs
@@ -13,7 +13,7 @@ namespace pTyping.Songs {
 		[JsonProperty]
 		public string TextToShow;
 		[JsonProperty]
-		public int Time;
+		public double Time;
 
 		/// <summary>
 		/// Whether or not the note has been hit
